third_party/blink/web_tests/fast/css/cue-serialize.html

<!DOCTYPE html>
<style>
::cue {}
::cue(c) {}
::cue(c, .class, #id:past, i:future) {}
</style>
<script src="../../resources/js-test.js"></script>
<script>
description("Check serialization of ::cue and ::cue()");

var rules = document.styleSheets[0].cssRules;

shouldBe("rules.length", "3");
shouldBe("rules[0].selectorText", "'::cue'");
shouldBe("rules[1].selectorText", "'::cue(c)'");
shouldBe("rules[2].selectorText", "'::cue(c, .class, #id:past, i:future)'");
</script>
