third_party/blink/web_tests/platform/generic/fast/css/cue-serialize-expected.txt

Check serialization of ::cue and ::cue()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS rules.length is 3
PASS rules[0].selectorText is '::cue'
PASS rules[1].selectorText is '::cue(c)'
PASS rules[2].selectorText is '::cue(c, .class, #id:past, i:future)'
PASS successfullyParsed is true

TEST COMPLETE
